+++ CHANGES.md
@@ -5,7 +5,7 @@
 -   Add tasks for submitting HTCondor DAGs.
 
 -   Add a new module, `gwcelery.tasks.lalinference`, which provides tasks to
-    start Parameter Estimation with LALInference and upload the results to
+    start parameter estimation with LALInference and upload the results to
     GraceDB.
 
 ## 0.1.4 (2018-10-29)