+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.1.5 (unreleased)
+## 0.1.5 (2018-11-13)
 
 -   Add tasks for submitting HTCondor DAGs.